requirements.txt
selenium
webdriver-manager
colorama
python-dotenv